requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
Automat==22.10.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
binance-connector==3.2.0
cached-property==1.5.2
celery==5.2.7
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
dateparser==1.2.0
Django==3.2.25
django-cors-headers==4.1.0
djangorestframework==3.15.1
flower==2.0.1
frozenlist==1.3.3
gunicorn==22.0.0
humanize==4.6.0
hyperlink==21.0.0
idna==3.7
importlib-metadata==4.8.1
incremental==22.10.0
kombu==5.2.3
multidict==6.0.5
numpy==1.21.6
packaging==24.0
prometheus-client==0.17.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycryptodome==3.20.0
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.1
redis==3.5.3
regex==2024.4.16
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tornado==6.2
Twisted==23.8.0
typing_extensions==4.7.1
tzlocal==5.1
ujson==5.7.0
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.6.1
websockets==11.0.3
yarl==1.9.4
zipp==3.15.0
zope.interface==6.4.post2
